Fix battle with multiple accounts
Two diferent accounts can not join the battle a user my ask for help in relation to this one made a whole series of tests entering the bot before then the phone one can join the battle if one goes with the phone then with the Bot the bot gives message. Could you do something with this bug?
if delete this ? it works or is required ?